requirements.txt
-e git+https://github.com/commonsense/conceptnet5@941cf7ec5d4c2761ab8d17a96c7616c388cc1e22#egg=ConceptNet-master
Flask==0.10.1
Flask-Cors==2.0.0
Flask-Limiter==0.7.9
ftfy==3.4.0
grako==3.5.1
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.3
limits==1.0.4
MarkupSafe==0.23
msgpack-python==0.4.6
nltk==3.0.2
Pillow==2.8.1
PyYAML==3.11
requests==2.6.0
six==1.9.0
uWSGI==2.0.10
Werkzeug==0.10.4
xmltodict==0.9.2
